Move gitlab export to backend/gitlab.ts
