+++ client/test/preview/unit/util/gitlabFactory.test.ts
@@ -1,4 +1,4 @@
-import GitlabInstance from 'model/backend/gitlab/interfaces';
+import GitlabInstance from 'model/backend/gitlab/gitlab';
 import createGitlabInstance from 'preview/util/gitlabFactory';
 
 describe('createGitlabInstance', () => {
